requirements.txt
asgi_redis==0.10.0
autobahn==0.13.1
channels==0.11.0
daphne==0.10.3
django<1.10
django-extensions==1.6.7
django-model-utils==2.5
gunicorn==19.6.0
ipython>=5.0.0b2,<=5.0.0
pytest==2.9.2
pytest-django==2.9.1
pylama==7.0.9
requests==2.10.0

# these requirements are installed in the base image
# and should not have pinned versions to prevent forced reinstallation
psycopg2
